chatbot_service/pyproject.toml
[tool.poetry]
name = "chatbot-service"
version = "0.1.0"
description = ""
authors = ["daria425 <darianaumova5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
pydantic = "^2.10.6"
requests = "^2.32.3"
uuid = "^1.30"
google-cloud-pubsub = "^2.28.0"
google-cloud-aiplatform = "^1.80.0"
motor = "^3.7.0"
fastapi = {extras = ["standard"], version = "^0.115.8"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
